pyproject.toml
[tool.poetry]
name = "my-app"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.2.0"
uWSGI = "^2.0.20"
dj-rest-auth = "^2.2.4"
django-cors-headers = "^3.13.0"
djangorestframework-api-key = "^2.2.0"
django-better-admin-arrayfield = "^1.4.2"
djangorestframework-simplejwt = "^5.2.0"
django-extensions = "^3.2.0"
django-debug-toolbar = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
